input: report A button, grip and thumbstick/trackpad to legacy input
Also disables reporting of the secondary Knuckles trackpad via the axis
type properties. This is done because when reporting this secondary
trackpad, Google Earth acts differently and I haven't been able
to figure out how to get it to properly read the thumbstick.

As a result, in Google Earth the Knuckles show up as Quest controllers,
but this seems to work fine (and actually be closer to what should be
correct behavior anyway, since they appear as Vive wands through
SteamVR).