compositor: avoid recreating swapchain if width and height are smaller
For some unknowable reason, Thumper submits a left and right eye with different
widths - one of them is off by one. This would cause the swapchain to
constantly be recreated in a loop before, since the size keeps changing
between eyes, and nothing would be displayed.

The easiest way of dealing with this is to just avoid recreating the swapchain
if it can just be reused. Hopefully games aren't submitting massive images
followed by much smaller ones.